all_tests.c: Implement lexer texts.
This included changing the behaviour of append_char in lib/lexer.c as the former version reallocated without changing our old pointer to the dynamic string. It happened to work because the test makes a small reallocation and the buffer has a high chance to stay in-place in those cases.

The error was caught by `scan-build make test`.

This finishes #2 for the most part.