Reorganized documentation.
Documentation for exported symbols now belongs to header files, while documentation for non-exported symbol remains in their original .c files. Part of this reorganization includes the fact that append_char() is no longer exported in lexer.h. To accomodate the fact that it is still desirable to test it, I changes the #include-s in all_tests.c to include the .c files. Several changes in the Makefile followed do fix errors generated by this change.

It was a big commit, although it adds no functionality