include/minunit.h: Add macros to be safer with tests.
This adds two macros, which are now used in tests: mu_assert_freeing() takes a third argument apart from the two of mu_assert. This new argument is a pointer that should be free()d before returning if the test fails.

mu_assert_running() is more general: the third argument should be a block that will run before returning if the test fails. Used to free several pointers if necessary.

The change in all_tests.c is simply the adoption of these macros.

Now tests are memory safe.